refactor: rename util to utils.
