+++ genvarloader/__init__.py
@@ -11,7 +11,7 @@ from .haplotypes import Haplotypes
 from .intervals import Intervals
 from .loader import GVL, construct_virtual_data
 from .torch import get_dataloader, get_sampler
-from .util import read_bedlike, with_length
+from .utils import read_bedlike, with_length
 from .variants import Variants
 from .zarr import ZarrTracks
