+++ genvarloader/fasta.py
@@ -9,7 +9,7 @@ from numpy.typing import ArrayLike, NDArray
 from typing_extensions import assert_never
 
 from .types import Reader
-from .util import get_rel_starts, normalize_contig_name
+from .utils import get_rel_starts, normalize_contig_name
 
 
 class NoPadError(Exception):
