+++ genvarloader/torch.py
@@ -8,7 +8,7 @@ from loguru import logger
 from numpy.typing import NDArray
 
 from .types import Reader
-from .util import construct_virtual_data, process_bed
+from .utils import construct_virtual_data, process_bed
 
 try:
     import torch
