+++ genvarloader/variants/records.py
@@ -12,7 +12,7 @@ from numpy.typing import ArrayLike, NDArray
 from tqdm.auto import tqdm
 from typing_extensions import Self
 
-from ..util import normalize_contig_name
+from ..utils import normalize_contig_name
 
 try:
     import cyvcf2
